Rename components to capitalized
