+++ src/components/LineChart/index.jsx
@@ -1,4 +1,3 @@
-// LineChart.js
 import React from "react";
 import { Line } from "react-chartjs-2";
 import {
