+++ src/screens/login/index.jsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import { useSelector } from "react-redux";
 import { Card, CardBody, Row, Button, Form, Spinner } from "reactstrap";
-import { Input } from "../../components/input";
-import { EyeIcon, EyeOffIcon } from "../../components/icons";
+import Input from "../../components/Input";
+import { EyeIcon, EyeOffIcon } from "../../components/Icons";
 import loginInputs from "../../config/constant/loginInputs";
 import useAuth from "../../hooks/useAuth";
 import useUser from "../../hooks/useUser";
