+++ src/screens/sign-up/index.jsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import { useSelector } from "react-redux";
 import { Button, Card, CardBody, Form, Row, Spinner } from "reactstrap";
-import { Input } from "../../components/input";
-import { EyeIcon, EyeOffIcon } from "../../components/icons";
+import Input from "../../components/Input";
+import { EyeIcon, EyeOffIcon } from "../../components/Icons";
 import signUpInputs from "../../config/constant/signUpInputs";
 import useAuth from "../../hooks/useAuth";
 import useUser from "../../hooks/useUser";
